+++ integration/src/test/java/org/kie/hacep/PodAsReplicaTest.java
@@ -1,15 +1,12 @@
 package org.kie.hacep;
 
 import java.util.ArrayList;
-import java.util.ConcurrentModificationException;
 import java.util.Iterator;
 import java.util.List;
 
 import org.apache.kafka.clients.consumer.ConsumerRecord;
 import org.apache.kafka.clients.consumer.ConsumerRecords;
 import org.apache.kafka.clients.consumer.KafkaConsumer;
-import org.junit.After;
-import org.junit.Before;
 import org.junit.Test;
 import org.kie.hacep.core.Bootstrap;
 import org.kie.hacep.core.infra.election.State;
@@ -17,7 +14,6 @@ import org.kie.hacep.message.ControlMessage;
 import org.kie.hacep.sample.kjar.StockTickEvent;
 import org.kie.remote.RemoteFactHandle;
 import org.kie.remote.RemoteKieSession;
-import org.kie.remote.TopicsConfig;
 import org.kie.remote.command.FireUntilHaltCommand;
 import org.kie.remote.command.InsertCommand;
 import org.kie.remote.command.RemoteCommand;
@@ -27,46 +23,19 @@ import org.slf4j.LoggerFactory;
 import static org.junit.Assert.*;
 import static org.kie.remote.util.SerializationUtil.deserialize;
 
-public class PodAsReplicaTest {
+public class PodAsReplicaTest extends KafkaFullTopicsTests {
 
-    private final String TEST_KAFKA_LOGGER_TOPIC = "logs";
-    private final String TEST_TOPIC = "test";
-    private KafkaUtilTest kafkaServerTest;
     private Logger logger = LoggerFactory.getLogger(PodAsReplicaTest.class);
-    private Logger kafkaLogger = LoggerFactory.getLogger("org.hacep");
-    private EnvConfig config;
-    private TopicsConfig topicsConfig;
-
-    @Before
-    public void setUp() throws Exception {
-        config = KafkaUtilTest.getEnvConfig();
-        topicsConfig = TopicsConfig.getDefaultTopicsConfig();
-        kafkaServerTest = new KafkaUtilTest();
-        kafkaServerTest.startServer();
-        kafkaServerTest.createTopics(TEST_KAFKA_LOGGER_TOPIC,
-                                     TEST_TOPIC,
-                                     config.getEventsTopicName(),
-                                     config.getControlTopicName(),
-                                     config.getSnapshotTopicName(),
-                                     config.getKieSessionInfosTopicName());
-    }
-
-    @After
-    public void tearDown() {
-        kafkaServerTest.tearDown();
-    }
-
-
 
     @Test(timeout = 60000L)
     public void processOneSentMessageAsLeaderAndThenReplicaTest() {
-        Bootstrap.startEngine(config);
+        Bootstrap.startEngine(envConfig);
         Bootstrap.getConsumerController().getCallback().updateStatus(State.LEADER);
         KafkaConsumer eventsConsumer = kafkaServerTest.getConsumer("",
-                                                                   config.getEventsTopicName(),
+                                                                   envConfig.getEventsTopicName(),
                                                                    Config.getConsumerConfig("eventsConsumerProcessOneSentMessageAsLeaderTest"));
         KafkaConsumer controlConsumer = kafkaServerTest.getConsumer("",
-                                                                    config.getControlTopicName(),
+                                                                    envConfig.getControlTopicName(),
                                                                     Config.getConsumerConfig("controlConsumerProcessOneSentMessageAsLeaderTest"));
 
         KafkaConsumer<byte[], String> kafkaLogConsumer = kafkaServerTest.getStringConsumer(TEST_KAFKA_LOGGER_TOPIC);
@@ -81,14 +50,14 @@ public class PodAsReplicaTest {
             Iterator<ConsumerRecord<String, byte[]>> eventsRecordIterator = eventsRecords.iterator();
 
             ConsumerRecord<String, byte[]> eventsRecord = eventsRecordIterator.next();
-            assertEquals(eventsRecord.topic(), config.getEventsTopicName());
+            assertEquals(eventsRecord.topic(), envConfig.getEventsTopicName());
             RemoteCommand remoteCommand = deserialize(eventsRecord.value());
             assertEquals(eventsRecord.offset(), 0);
             assertNotNull(remoteCommand.getId());
             assertTrue( remoteCommand instanceof FireUntilHaltCommand );
 
             eventsRecord = eventsRecordIterator.next();
-            assertEquals(eventsRecord.topic(), config.getEventsTopicName());
+            assertEquals(eventsRecord.topic(), envConfig.getEventsTopicName());
             remoteCommand = deserialize(eventsRecord.value());
             assertEquals(eventsRecord.offset(), 1);
             assertNotNull(remoteCommand.getId());
@@ -171,14 +140,14 @@ public class PodAsReplicaTest {
 
     private void checkFireSideEffects( ConsumerRecord<String, byte[]> controlRecord ) {
         // FireUntilHalt command has no side effects
-        assertEquals(controlRecord.topic(), config.getControlTopicName());
+        assertEquals(controlRecord.topic(), envConfig.getControlTopicName());
         ControlMessage controlMessage = deserialize(controlRecord.value());
         assertEquals(controlRecord.offset(), 0);
         assertTrue(controlMessage.getSideEffects().isEmpty());
     }
 
     private void checkInsertSideEffects( ConsumerRecord<String, byte[]> eventsRecord, ConsumerRecord<String, byte[]> controlRecord ) {
-        assertEquals(controlRecord.topic(), config.getControlTopicName());
+        assertEquals(controlRecord.topic(), envConfig.getControlTopicName());
         ControlMessage controlMessage = deserialize(controlRecord.value());
         assertEquals(controlRecord.offset(), 1);
         assertTrue(!controlMessage.getSideEffects().isEmpty());
